Replace permissions/index.js by conditionals.js
